RFC 9022: bug fixing Ulf
git-svn-id: http://10.205.116.236/svn/AIR@345 d2170373-ae6f-44b7-9150-2eec3990abfc

+++ branches/2.00/src/com/bayerbbs/applrepos/hibernate/CiEntitiesHbn.java
@@ -6,7 +6,6 @@ import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.sql.Statement;
 import java.util.ArrayList;
-import java.util.Collection;
 import java.util.Collections;
 import java.util.List;
 
@@ -767,7 +766,7 @@ public class CiEntitiesHbn {
 		return output;
 	}
 	
-	public static void saveCiRelations(Long tableId, Long ciId, String ciRelationsAddList, String ciRelationsDeleteList, String direction, String cwid) {
+	public static void saveCiRelations(Integer tableId, Long ciId, String ciRelationsAddList, String ciRelationsDeleteList, String direction, String cwid) {
 		String sql = "{call pck_air.p_save_relations(?,?,?,?,?,?)}";//"begin pck_air.p_save_relations(?,?,?,?,?,?);//"EXEC pck_air.p_save_relations ("+tableId+", "+ciId+", "+ciRelationsAddList+", "+ciRelationsDeleteList+", "+direction+", "+cwid+")";
 		
 		Transaction ta = null;
